requirements.txt
openpyxl>=3.1.0
watchdog>=3.0.0
psutil>=5.9.0
wcwidth>=0.2.0
lz4>=4.3.0
zstandard>=0.22.0
polars>=0.20.0
xlsx2csv>=0.8.0
Flask>=3.0.0
GitPython>=3.1.40